refactor: refactoring page-header
